requirements.txt
python-dotenv==1.0.1
requests== 2.32.3
aiogram==3.8.0
